update secret optional

+++ test/e2e/scaffold/scaffold.go
@@ -457,28 +457,28 @@ func (s *Scaffold) beforeEach() {
 func (s *Scaffold) afterEach() {
 	defer ginkgo.GinkgoRecover()
 
-	if ginkgo.CurrentSpecReport().Failed() {
+	if ginkgo.CurrentSpecReport().Failed() && os.Getenv("E2E_ENV") != "dev" {
+		// dump and delete related resource
+		_, _ = fmt.Fprintln(ginkgo.GinkgoWriter, "Dumping namespace contents")
+		output, _ := k8s.RunKubectlAndGetOutputE(ginkgo.GinkgoT(), s.kubectlOptions, "get", "deploy,sts,svc,pods")
+		if output != "" {
+			_, _ = fmt.Fprintln(ginkgo.GinkgoWriter, output)
+		}
+		output, _ = k8s.RunKubectlAndGetOutputE(ginkgo.GinkgoT(), s.kubectlOptions, "describe", "pods")
+		if output != "" {
+			_, _ = fmt.Fprintln(ginkgo.GinkgoWriter, output)
+		}
+		// Get the logs of apisix
+		output = s.GetDeploymentLogs("apisix-deployment-e2e-test")
+		if output != "" {
+			_, _ = fmt.Fprintln(ginkgo.GinkgoWriter, output)
+		}
+		// Get the logs of ingress
+		output = s.GetDeploymentLogs("ingress-apisix-controller-deployment-e2e-test")
+		if output != "" {
+			_, _ = fmt.Fprintln(ginkgo.GinkgoWriter, output)
+		}
 		if os.Getenv("E2E_ENV") == "ci" {
-			// dump and delete related resource
-			_, _ = fmt.Fprintln(ginkgo.GinkgoWriter, "Dumping namespace contents")
-			output, _ := k8s.RunKubectlAndGetOutputE(ginkgo.GinkgoT(), s.kubectlOptions, "get", "deploy,sts,svc,pods")
-			if output != "" {
-				_, _ = fmt.Fprintln(ginkgo.GinkgoWriter, output)
-			}
-			output, _ = k8s.RunKubectlAndGetOutputE(ginkgo.GinkgoT(), s.kubectlOptions, "describe", "pods")
-			if output != "" {
-				_, _ = fmt.Fprintln(ginkgo.GinkgoWriter, output)
-			}
-			// Get the logs of apisix
-			output = s.GetDeploymentLogs("apisix-deployment-e2e-test")
-			if output != "" {
-				_, _ = fmt.Fprintln(ginkgo.GinkgoWriter, output)
-			}
-			// Get the logs of ingress
-			output = s.GetDeploymentLogs("ingress-apisix-controller-deployment-e2e-test")
-			if output != "" {
-				_, _ = fmt.Fprintln(ginkgo.GinkgoWriter, output)
-			}
 			err := k8s.DeleteNamespaceE(s.t, s.kubectlOptions, s.namespace)
 			assert.Nilf(ginkgo.GinkgoT(), err, "deleting namespace %s", s.namespace)
 		}
